refactor(shared/api): http error 관련 파일들 https/error/ 하위로 이동
