+++ src/app/_providers/react-query.provider.tsx
@@ -3,8 +3,8 @@ import { ReactNode } from 'react';
 import { MutationCache, QueryCache, QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { ReactQueryStreamedHydration } from '@tanstack/react-query-next-experimental';
-import { getErrorMessage } from '@/shared/api/get-error-message';
-import isAuthError from '@/shared/api/is-auth-error';
+import { getErrorMessage } from '@/shared/api/http/error/get-error-message';
+import isAuthError from '@/shared/api/http/error/is-auth-error';
 import { FIVE_MINUTES } from '@/shared/config/time';
 import { Dialog } from '@/shared/ui/components/dialog';
 import { Typography } from '@/shared/ui/components/typography';
