+++ src/app/parties/(lobby)/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { PartyroomCreateCard } from '@/features/partyroom/create';
 import { MainPartyroomCard, PartyroomList } from '@/features/partyroom/list';
-import SuspenseWithErrorBoundary from '@/shared/api/suspense-with-error-boundary.component';
+import SuspenseWithErrorBoundary from '@/shared/api/http/error/suspense-with-error-boundary.component';
 import { cn } from '@/shared/lib/functions/cn';
 import { useAppRouter } from '@/shared/lib/router/use-app-router.hook';
 import { Sidebar } from '@/widgets/sidebar';
