+++ src/shared/api/http/error/is-auth-error.ts
