+++ src/shared/api/http/error/suspense-with-error-boundary.component.tsx
